+++ src/static/plugins/tb_cpu_usage/plugin/template.js
@@ -1,0 +1,4 @@
+const totalUsage = cores.reduce((total, core) => total + core.usage, 0);
+const used = totalUsage / cores.length;
+
+return [icon("LuCpu"), " ", used.toFixed(0) + "%"];