+++ src/static/plugins/tb_memory_usage/plugin/template.js
@@ -1,0 +1,17 @@
+const units = ["KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"];
+
+let total = memory.total;
+let used = memory.total - memory.free;
+let unit = "B";
+
+units.forEach((unitSize) => {
+  if (total >= 1024) {
+    total /= 1024;
+    used /= 1024;
+    unit = unitSize;
+  } else {
+    return;
+  }
+});
+
+return [icon("FaMemory"), " ", used.toFixed(0) + unit + " / " + total.toFixed(0) + unit];